layout clean up
